+++ index.js
@@ -4,6 +4,8 @@ const { RNCloverBridge } = NativeModules;
 
 export default {
   ...RNCloverBridge,
+  authenticate: (force = false, timeout = 10000) =>
+    RNCloverBridge.authenticate(force, timeout),
   isFlex: () => RNCloverBridge.isFlex,
   isMini: () => RNCloverBridge.isMini,
   getSpaVersion: () => RNCloverBridge.getSpaVersion,